Move utils to corresponding chain folder
