+++ src/chains/EVM/types.ts
@@ -1,5 +1,10 @@
 import type * as ethers from 'ethers'
-import { type ChainProvider, type NearAuthentication } from '../types'
+import {
+  type ChainSignatureContracts,
+  type NearNetworkIds,
+  type ChainProvider,
+  type NearAuthentication,
+} from '../types'
 import { type KeyDerivationPath } from '../../kdf/types'
 
 export type EVMTransaction = ethers.TransactionLike
@@ -12,4 +17,10 @@ export interface EVMRequest {
   nearAuthentication: NearAuthentication
   fastAuthRelayerUrl?: string
   derivationPath: KeyDerivationPath
+}
+export interface FetchEVMAddressRequest {
+  signerId: string
+  path: KeyDerivationPath
+  nearNetworkId: NearNetworkIds
+  multichainContractId: ChainSignatureContracts
 }
